Refactor: create modular file structure
