+++ src/movieDescription/MovieDescription.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Link } from 'react-router-dom';
 import './MovieDescription.css';
-import thinking from './thinking.gif'
+import thinking from '../media/thinking.gif'
 
 
 class MovieDescription extends Component {
